Rewrite crud - Delete is now secure

+++ api/login/delete.php
@@ -6,6 +6,15 @@ header("Access-Control-Allow-Methods: POST");
 header("Access-Control-Max-Age: 3600");
 header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
 
+// required to encode json web token
+include_once 'core.php';
+include_once '../../libs/php-jwt-master/src/BeforeValidException.php';
+include_once '../../libs/php-jwt-master/src/ExpiredException.php';
+include_once '../../libs/php-jwt-master/src/SignatureInvalidException.php';
+include_once '../../libs/php-jwt-master/src/JWT.php';
+use \Firebase\JWT\JWT;
+
+
 // include database and object file
 include_once '../groepen/database.php';
 include_once '../objects/user.php';
@@ -20,26 +29,57 @@ $user = new User($db);
 // get product id
 $data = json_decode(file_get_contents("php://input"));
 
+$jwt=isset($data->jwt) ? $data->jwt : "";
+
+if($jwt){
+
+    // if decode succeed, show user details
+    try {
+
+        // decode jwt
+        $decoded = JWT::decode($jwt, $key, array('HS256'));
+
 // set product id to be deleted
-$user->id = $data->id;
+        $user->id = $data->id;
 
 // delete the product
-if($user->delete()){
+        if ($user->delete()) {
 
-    // set response code - 200 ok
-    http_response_code(200);
+            // set response code - 200 ok
+            http_response_code(200);
 
-    // tell the user
-    echo json_encode(array("message" => "User was deleted."));
+            // tell the user
+            echo json_encode(array("message" => "User was deleted."));
+        } // if unable to delete the product
+        else {
+
+            // set response code - 503 service unavailable
+            http_response_code(503);
+
+            // tell the user
+            echo json_encode(array("message" => "Unable to delete User."));
+        }
+    }
+        // if decode fails, it means jwt is invalid
+    catch (Exception $e){
+
+            // set response code
+            http_response_code(401);
+
+            // show error message
+            echo json_encode(array(
+                "message" => "Access denied.",
+                "error" => $e->getMessage()
+            ));
+        }
 }
-
-// if unable to delete the product
+// show error message if jwt is empty
 else{
 
-    // set response code - 503 service unavailable
-    http_response_code(503);
+    // set response code
+    http_response_code(401);
 
-    // tell the user
-    echo json_encode(array("message" => "Unable to delete User."));
+    // tell the user access denied
+    echo json_encode(array("message" => "Access denied."));
 }
 ?>
